Prepare main node script

+++ include/point_cloud_accumulator_pkg/filters/color_consistency_filter.hpp
@@ -11,10 +11,10 @@ namespace point_cloud_accumulator_pkg::filters
   class ColorConsistencyFilter : public Filter
   {
     public:
-      ColorConsistencyFilter(size_t history_size = 10, uint8_t saturation_threshold = 240);
+      ColorConsistencyFilter(int history_size = 10, uint8_t saturation_threshold = 240);
 
     protected:
-      CloudPtr applyFilter(const CloudPtr& cloud) const override;
+      CloudPtr applyFilter(const CloudPtr &cloud) const override;
 
     private:
       bool isOverexposed(const PointT &pt) const;
@@ -23,7 +23,7 @@ namespace point_cloud_accumulator_pkg::filters
       float rollingAverage() const;
 
       mutable std::deque<float> brightness_history_;
-      size_t history_size_;
+      int history_size_;
       uint8_t saturation_threshold_;
   };
