+++ include/point_cloud_accumulator_pkg/filters/temporal_filter.hpp
@@ -15,7 +15,7 @@ namespace point_cloud_accumulator_pkg::filters
   class TemporalFilter : public Filter
   {
     public:
-      TemporalFilter(size_t history_size = 5, float distance_thr_m = 0.02f, float min_appearance_ratio = 0.6f);
+      TemporalFilter(int history_size = 5, float distance_thr_m = 0.02f, float min_appearance_ratio = 0.6f);
       void setDistanceThreshold(float distance_thr);
 
     protected:
@@ -27,7 +27,7 @@ namespace point_cloud_accumulator_pkg::filters
       /** Double ended queue keeping a limited history of point cloud frames. */
       mutable std::deque<CloudPtr> frame_history_;
       /** Size limit of the point cloud frame history. */
-      size_t history_size_;
+      int history_size_;
       /** Distance threshold, in meters, for a point to be considered the same across frames. */
       float distance_thr_m_;
       /** Frequency (as a ratio) of occurences for a point to be persisted. */
