+++ src/filters/tf_outlier_filter.cpp
@@ -3,7 +3,7 @@
 namespace point_cloud_accumulator_pkg::filters
 {
 
-  TFOutlierFilter::TFOutlierFilter(double max_translation_m, double max_rotation_deg, uint8_t history_size)
+  TFOutlierFilter::TFOutlierFilter(double max_translation_m, double max_rotation_deg, int history_size)
     : max_translation_m_(max_translation_m)
     , max_rotation_deg_(max_rotation_deg)
     , history_size_(history_size)
